Fix imports after moving Dashboard
